:sparkles: finalizando projeto

+++ src/main/java/com/gerenciador/dashboard/record/TopCincoRecord.java
@@ -2,5 +2,6 @@ package com.gerenciador.dashboard.record;
 
 import java.math.BigDecimal;
 
-public record TopCincoRecord(String usuario, Integer totalPedidos, BigDecimal totalGasto) {
+
+public record TopCincoRecord(String usuario, Long totalPedidos, BigDecimal totalGasto) {
 }
